refactor: migrate README check to turbo-native per-package pattern
Replaces the workspace-root `.mjs` scanner with an internal
`@glion/check-readme` package that every `@glion/*` package consumes
via a `check:readme` script orchestrated by Turborepo. Matches the
repo's established pattern (`build`, `test`, `check-types` all work
the same way) and gives future per-package checks a natural home.

What changes:

- Add `packages/check-readme/` — private internal tool providing the
  `glion-check-readme` bin. Zero runtime dependencies; 8 unit tests
  via Node's built-in `node:test` runner.
- Add `check:readme` task to `turbo.json` with `package.json` and
  `README.md` as inputs so Turbo caches unchanged packages.
- Add `check:readme` script and `@glion/check-readme` workspace
  devDependency to every public package.json (41 packages).
- Add `check:readme` script and `@glion/check-readme` devDep to the
  turbo generator template (`turbo/generators/templates/package.json.hbs`)
  so new packages scaffolded via `turbo gen package` automatically
  participate.
- Rewrite the root `check:readmes` script to invoke
  `turbo run check:readme` (same public command name; Turbo dispatches
  under the hood).
- Drop `check:readmes:test` from the root — the tool's tests now run
  as part of `pnpm test` via Turbo's standard `test` task.
- Remove `tools/check-package-readmes.mjs` and its test file.

Benefits:

- Turbo caching: packages whose `README.md` and `package.json` haven't
  changed since the last successful check are skipped.
- Parallel execution: 41 packages checked concurrently (~1.3s local).
- Per-package filtering: `pnpm turbo run check:readme --filter=@glion/mllp`
  works out of the box.
- Consistent with repo conventions — future per-package lints (link
  validity, analogy scan, etc.) slot in the same way.
- New packages from `turbo gen package` automatically include the
  script + devDep — no contributor-copy-paste step.

Transitional state: `pnpm check:readmes` still fails on the current
repo (41 packages fail, matching the expected baseline from the prior
`.mjs` scanner). Phase 2 will write the 16 missing READMEs; Phase 3
will rewrite the 25 existing to match the template; Phase 4 removes
`continue-on-error: true` from the CI advisory step.

Extends PR #592.
