+++ package.json
@@ -25,8 +25,7 @@
     "build": "turbo run build",
     "check": "ultracite check",
     "check-types": "turbo run check-types",
-    "check:readmes": "node tools/check-package-readmes.mjs",
-    "check:readmes:test": "node --test tools/check-package-readmes.test.mjs",
+    "check:readmes": "turbo run check:readme",
     "ci:install": "pnpm install --no-frozen-lockfile",
     "ci:publish": "pnpm recursive publish --no-git-checks --access public",
     "ci:update": "pnpm --recursive update",
