+++ packages/ack/package.json
@@ -36,6 +36,7 @@
   "scripts": {
     "build": "tsdown && tsc --emitDeclarationOnly",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
@@ -48,6 +49,7 @@
     "nanoid": "5.1.7"
   },
   "devDependencies": {
+    "@glion/check-readme": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/to-hl7v2": "workspace:*",
     "@glion/tsconfig": "workspace:*",
