+++ packages/annotate-delimiters/package.json
@@ -18,6 +18,7 @@
   "scripts": {
     "build": "tsup && tsc --project tsconfig.json",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:watch": "vitest watch"
   },
@@ -27,6 +28,7 @@
   "devDependencies": {
     "@glion/ast": "workspace:*",
     "@glion/builder": "workspace:*",
+    "@glion/check-readme": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/tsconfig": "workspace:*",
     "@glion/utils": "workspace:*",
