+++ packages/ast/package.json
@@ -28,9 +28,11 @@
     "access": "public"
   },
   "scripts": {
-    "check-types": "tsc ./index.d.ts --noEmit"
+    "check-types": "tsc ./index.d.ts --noEmit",
+    "check:readme": "glion-check-readme"
   },
   "devDependencies": {
+    "@glion/check-readme": "workspace:*",
     "@glion/tsconfig": "workspace:*",
     "@types/node": "^25.5.0",
     "@types/unist": "^3.0.3",
