+++ packages/check-readme/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@glion/check-readme",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Internal tool: validate @glion/* package READMEs against the template contract.",
+  "bin": {
+    "glion-check-readme": "./src/cli.mjs"
+  },
+  "type": "module",
+  "scripts": {
+    "test": "node --test test/*.test.mjs"
+  },
+  "engines": {
+    "node": ">=18"
+  }
+}
