+++ packages/glion/package.json
@@ -46,6 +46,7 @@
   "scripts": {
     "build": "tsdown && tsc -p tsconfig.build.json --emitDeclarationOnly",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:watch": "vitest"
   },
@@ -55,6 +56,7 @@
     "zod": "^3.23.0"
   },
   "devDependencies": {
+    "@glion/check-readme": "workspace:*",
     "@glion/hl7v2": "workspace:*",
     "@glion/mllp-ack": "workspace:*",
     "@glion/testing": "workspace:*",
