+++ packages/lint-max-message-size/package.json
@@ -38,6 +38,7 @@
   "scripts": {
     "build": "tsdown && tsc --emitDeclarationOnly",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
@@ -52,6 +53,7 @@
   "devDependencies": {
     "@glion/ast": "workspace:*",
     "@glion/builder": "workspace:*",
+    "@glion/check-readme": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/tsconfig": "workspace:*",
     "@types/node": "^25.5.0",
