+++ packages/lint-profile-extra-components/package.json
@@ -35,6 +35,7 @@
   "scripts": {
     "build": "tsdown && tsc --emitDeclarationOnly",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "bench": "vitest bench --run",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
@@ -49,6 +50,7 @@
   "devDependencies": {
     "@glion/annotate-profile-context": "workspace:*",
     "@glion/builder": "workspace:*",
+    "@glion/check-readme": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/tsconfig": "workspace:*",
     "@types/node": "^25.5.0",
