+++ packages/preset-lint-recommended/package.json
@@ -38,6 +38,7 @@
   "scripts": {
     "build": "tsdown && tsc --emitDeclarationOnly",
     "check-types": "tsc --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
@@ -53,6 +54,7 @@
   "devDependencies": {
     "@glion/ast": "workspace:*",
     "@glion/builder": "workspace:*",
+    "@glion/check-readme": "workspace:*",
     "@glion/jsonify": "workspace:*",
     "@glion/parser": "workspace:*",
     "@glion/testing": "workspace:*",
