+++ packages/profiles/package.json
@@ -43,6 +43,7 @@
     "bench": "vitest bench --run",
     "build": "tsdown && tsc -p tsconfig.build.json --emitDeclarationOnly",
     "check-types": "tsc -p tsconfig.build.json --noEmit",
+    "check:readme": "glion-check-readme",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
@@ -54,6 +55,7 @@
   },
   "devDependencies": {
     "@glion/ast": "workspace:*",
+    "@glion/check-readme": "workspace:*",
     "@glion/config": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/tsconfig": "workspace:*",
