+++ turbo.json
@@ -23,6 +23,10 @@
     },
     "test": {
       "dependsOn": ["^build"]
+    },
+    "check:readme": {
+      "inputs": ["package.json", "README.md"],
+      "outputs": []
     }
   }
 }
